+++ CHANGELOG.md
@@ -52,7 +52,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Bugfixes
 
-- [#334](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/334) fix: add missing sudo export
+- [#334](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/334) fix:
+  add missing sudo export
+- [#340](https://github.com/babylonlabs-io/cosmos-bsn-contracts/pull/340) fix:
+  delegation natural expiry
 
 ## [v0.17.0](https://github.com/babylonlabs-io/cosmos-bsn-contracts/tree/v0.17.0) (2025-08-08)